feat: create Operator struct with new and start
