+++ Cargo.toml
@@ -22,7 +22,6 @@ members = [
     "crates/utils/",
     "crates/nodeapi/",
     "crates/eigensdk/",
-    "examples/aggregator/",
     "examples/anvil-utils",
     "examples/avsregistry-read",
     "examples/avsregistry-write",
